backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
spotipy
httpx
python-multipart
python-dotenv
bcrypt
email-validator>=2.0.0
fastapi-cache2
ujson